requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dashscope==1.20.4
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
et-xmlfile==2.0.0
faiss-cpu==1.8.0.post1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gradio==5.29.0
gradio-client==1.10.0
greenlet==3.2.1
griffe==1.7.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llama-index-core==0.10.67
llama-index-embeddings-dashscope==0.1.4
llama-index-embeddings-openai==0.3.1
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
llama-index-readers-file==0.1.33
llama-index-vector-stores-faiss==0.1.2
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.3
mcp==1.7.0
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.1.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.55.3
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
propcache==0.3.1
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.3
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.8
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
